v2.4.8.4 -> v2.4.9
  - David Miller: sparc updates, FAT fs fixes, btaudio build fix
  - David Gibson: Orinoco driver update
  - Kevin Fleming: more disks the HPT controller doesn't like
  - David Miller: "min()/max()" cleanups. Understands signs and sizes.
  - Ben LaHaise: make vma merging more generous, help Mozilla /proc/<>/maps
  - Jens Axboe: CD updates
  - Trond Myklebust: save away NFS credentials in inode, so that mmap can
  writeout.
  - Mark Hemment: HIGHMEM ops cleanups
  - Jes Sorensen: use "unsigned long" for flags in various drivers

+++ drivers/i2o/i2o_block.c
@@ -713,7 +713,7 @@ static void i2o_block_reply(struct i2o_handler *h, struct i2o_controller *c, str
 static int i2ob_evt(void *dummy)
 {
 	unsigned int evt;
-	unsigned int flags;
+	unsigned long flags;
 	int unit;
 	int i;
 	//The only event that has data is the SCSI_SMART event.
@@ -1698,7 +1698,7 @@ void i2ob_del_device(struct i2o_controller *c, struct i2o_device *d)
 {	
 	int unit = 0;
 	int i = 0;
-	int flags;
+	unsigned long flags;
 
 	spin_lock_irqsave(&io_request_lock, flags);
